v6.2.5

+++ bus-core/src/main/java/org/aoju/bus/core/lang/mutable/MutableByte.java
@@ -32,7 +32,7 @@ import org.aoju.bus.core.toolkit.MathKit;
  * 可变 <code>byte</code> 类型
  *
  * @author Kimi Liu
- * @version 6.2.3
+ * @version 6.2.5
  * @see Byte
  * @since JDK 1.8+
  */
@@ -79,7 +79,7 @@ public class MutableByte extends Number implements Comparable<MutableByte>, Muta
 
     @Override
     public Byte get() {
-        return Byte.valueOf(this.value);
+        return this.value;
     }
 
     /**
